New uniform_int internal module
